requirements.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
bokeh==2.2.1
boto==2.49.0
boto3==1.14.24
botocore==1.17.24
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorcet==2.0.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dask==2.26.0
-e git+git@github.com:datapop-upf/dataPreprocessing.git@ffc5f6af689ca81633eaa7dd46bfa5d89238e14c#egg=dataPreprocessing
datashader==0.11.1
datashape==0.5.2
decorator==4.4.2
defusedxml==0.6.0
distributed==2.26.0
docutils==0.15.2
entrypoints==0.3
fasttext==0.9.2
filelock==3.0.12
fsspec==0.8.2
funcy==1.14
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
hdbscan==0.8.26
HeapDict==1.0.1
holoviews==1.13.3
httplib2==0.18.1
idna==2.10
iniconfig==1.0.1
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.34.0
locket==0.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
more-itertools==8.5.0
msgpack==1.0.0
MulticoreTSNE==0.1
multipledispatch==0.6.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.0.3
numba==0.51.2
numexpr==2.7.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
panel==0.9.7
param==1.9.3
parso==0.7.0
partd==1.1.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycparser==2.20
pyct==0.4.6
Pygments==2.6.1
pyLDAvis==2.1.2
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyviz-comms==0.7.6
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sentence-transformers==0.3.5
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
smart-open==2.1.0
sortedcontainers==2.2.2
spacy==2.3.2
srsly==1.0.2
statsmodels==0.11.1
tblib==1.7.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.8.1rc1
toml==0.10.1
toolz==0.10.0
torch==1.6.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
transformers==3.0.2
-e git+git@github.com:Oliph/twitterAccess.git@52fd1b7a8fef8aedf4aecd296b613a1d270ba9a2#egg=twitterAccess
typing-extensions==3.7.4.3
umap-learn==0.4.6
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
wrapt==1.12.1
xarray==0.16.0
zict==2.0.0
